refactor: shared resource utils
