+++ shared/package.json
@@ -9,7 +9,7 @@
     "dev": "nodemon -e ts -w ./src -x npm run build"
   },
   "devDependencies": {
-    "@types/node": "^18.11.9",
+    "@types/node": "18.0.0",
     "rimraf": "3.0.2",
     "typescript": "4.8.4"
   },
@@ -21,14 +21,14 @@
     "@monokle/validation": "0.7.1",
     "@open-policy-agent/opa-wasm": "1.8.0",
     "@reduxjs/toolkit": "1.9.0",
-    "electron": "^21.2.3",
-    "electron-log": "^4.4.8",
+    "electron": "21.2.3",
+    "electron-log": "4.4.8",
     "electron-store": "^8.1.0",
     "lodash": "4.17.21",
     "nodemon": "2.0.20",
     "react": "18.2.0",
     "react-monaco-editor": "0.50.1",
-    "reselect": "^4.1.7",
+    "reselect": "4.1.7",
     "runtypes": "6.6.0",
     "type-fest": "3.2.0",
     "yaml": "2.1.3"
