+++ shared/src/utils/env.ts
@@ -58,3 +58,16 @@ export function shellEnvSync(): any {
     }
   }
 }
+
+let mainProcessEnv: any | undefined;
+
+export function getMainProcessEnv() {
+  if (!mainProcessEnv) {
+    mainProcessEnv = shellEnvSync();
+  }
+  return mainProcessEnv;
+}
+
+export function setMainProcessEnv(env: any) {
+  mainProcessEnv = env;
+}
