+++ shared/src/utils/index.ts
@@ -3,5 +3,6 @@ export * from './env';
 export * from './fileSystem';
 export * from './hotkey';
 export * from './logs';
+export * from './resource';
 export * from './selectors';
 export * from './shell';
