+++ shared/src/utils/resource.ts
@@ -1,14 +1,16 @@
-import fs from 'fs';
+import 'electron';
+
+import {existsSync, readFileSync} from 'fs';
 import path from 'path';
 
-import {getMainProcessEnv} from '@utils/env';
+import {getMainProcessEnv} from './env';
 
 /**
  * Gets the absolute path to a statically bundled resource in the /resources folder
  */
-
 export function getStaticResourcePath(resourcePath: string) {
   const mainProcessEnv = getMainProcessEnv();
+
   if (mainProcessEnv?.NODE_ENV === 'test') {
     return path.join('resources', resourcePath);
   }
@@ -21,13 +23,12 @@ export function getStaticResourcePath(resourcePath: string) {
 /**
  * Loads the static resource at the specified relative path to /resources
  */
-
 export function loadResource(resourcePath: string) {
   const staticResourcePath = getStaticResourcePath(resourcePath);
-  return fs.existsSync(staticResourcePath) ? fs.readFileSync(staticResourcePath, 'utf8') : undefined;
+  return existsSync(staticResourcePath) ? readFileSync(staticResourcePath, 'utf8') : undefined;
 }
 
 export function loadBinaryResource(resourcePath: string): ArrayBuffer | undefined {
   const staticResourcePath = getStaticResourcePath(resourcePath);
-  return fs.existsSync(staticResourcePath) ? fs.readFileSync(staticResourcePath) : undefined;
+  return existsSync(staticResourcePath) ? readFileSync(staticResourcePath) : undefined;
 }
